Wanted to check if I can redirect to external website. Yes I can.
Reason is that I want to redirect the customer to PayPal page from backend,
this prevents customer from modifying payment in the frontend using developer tools of browser.

So, the whole payment pricing is processed in the backend.
(Still need to actually test it but for now I think it's possible).